--- src/components/sign-up/corporateMember.module.scss
@@ -1,7 +1,0 @@
-@import './commonSignup.module.scss';
-
-.textBlue {
-  & > strong {
-    color: #00a5e1;
-  }
-}
