--- src/components/sign-up/individualMember.module.scss
@@ -1,1 +1,0 @@
-@import './commonSignup.module.scss';
